requirements.txt
torch
torchvision
facenet-pytorch
Pillow
gradio